Kręci śmigłem, trzeba co 500ms cana = rtos potrzebny

+++ STM32/uart_can/Core/Inc/can.h
@@ -2,10 +2,11 @@
 #define INC_CAN_H_
 
 #include "main.h"
+#include <stdint.h>
 
 void FDCAN_Config(FDCAN_HandleTypeDef *hfdcan);
 void CAN_TransmitOverCan(FDCAN_HandleTypeDef *hfdcan, uint32_t id, uint8_t *data, uint8_t len);
 
-void CAN_TransmitVescCommand(FDCAN_HandleTypeDef *hfdcan, VESC_Id_t vescID, VESC_Command_t command, float value);
+void CAN_TransmitVescCommand(FDCAN_HandleTypeDef *hfdcan, VESC_Id_t vescID, VESC_Command_t command, uint16_t value);
 
 #endif /* INC_CAN_H_ */
